web: extract redesign-toggle-decorator from redesign-toggle-toolbar
