+++ client/storybook/src/preview.ts
@@ -9,12 +9,13 @@ import { withDesign } from 'storybook-addon-designs'
 import { setLinkComponent, AnchorLink } from '@sourcegraph/shared/src/components/Link'
 import { getIsRedesignEnabled, REDESIGN_CLASS_NAME } from '@sourcegraph/shared/src/util/useRedesignToggle'
 
+import { RedesignToggleDecorator } from './RedesignToggleDecorator'
 import { themeDark, themeLight, THEME_DARK_CLASS, THEME_LIGHT_CLASS } from './themes'
 
 const withConsoleDecorator: DecoratorFunction<ReactElement> = (storyFn, context): ReactElement =>
     withConsole()(storyFn)(context)
 
-export const decorators = [withDesign, withConsoleDecorator]
+export const decorators = [withDesign, withConsoleDecorator, RedesignToggleDecorator]
 
 export const parameters = {
     darkMode: {
